+++ csharpsrc/Tests/LeetCode/No16/Constraints.cs
@@ -1,0 +1,42 @@
+﻿// <copyright file="Constraints.cs" company="Ocean">
+// Copyright (c) Ocean. All rights reserved.
+// </copyright>
+
+namespace Tests.LeetCode.No16
+{
+    /// <summary>
+    /// The constraints of LeetCode No.16 question: Three Sum Closest.
+    /// </summary>
+    internal static class Constraints
+    {
+        /// <summary>
+        /// The min length of test number array.
+        /// </summary>
+        public const int MinLength = 3;
+
+        /// <summary>
+        /// The max length of test number array.
+        /// </summary>
+        public const int MaxLength = 500;
+
+        /// <summary>
+        /// The min value of element in test number array.
+        /// </summary>
+        public const int MinElementValue = -1000;
+
+        /// <summary>
+        /// The max value of element in test number array.
+        /// </summary>
+        public const int MaxElementValue = 1000;
+
+        /// <summary>
+        /// The min value of test target.
+        /// </summary>
+        public const int MinTargetValue = -10000;
+
+        /// <summary>
+        /// The max value of test target.
+        /// </summary>
+        public const int MaxTargetValue = 10000;
+    }
+}