+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Data Structures and Algorithms Training Change Log
 ### The change log of this repository.
 
+* [03/25/2022] Implement the LeetCode No.16 question: [Three Sum Closest](leetcode/questions/Question16.md) by C#.
+* [03/25/2022] Implement the LeetCode No.18 question: [Four Sum](leetcode/questions/Question18.md) by C#.
 * [03/25/2022] Implement the LeetCode No.15 question: [Three Sum](leetcode/questions/Question15.md) by C#.
 * [03/25/2022] Implement the LeetCode No.1 question: [Two Sum](leetcode/questions/Question1.md) by C#.
 * [03/22/2022] Initialize the C# solution and projects.